Configure Publish on Tag (#4)
* Refactor Python publishing workflow

- Updated the checkout action to version 3.5.3
- Set up Python 3.8 using actions/setup-python@v3.1.4
- Cached the installation of Poetry using actions/cache@v3.3.1
- Installed Poetry version 1.2.1 with snok/install-poetry@v1
- Used JRubics/poetry-publish@v1.17 to build and publish the Python package

* feat: Add testpypi repository for publishing

This commit adds the necessary configuration to publish packages to the testpypi repository. The `repository_name` is set to "testpypi" and the `repository_url` is set to "https://test.pypi.org/legacy/".

* feat: Add publishing workflow for Python package

This commit adds a new workflow `poetry-pytest.yml` that sets up the environment, installs Poetry, and publishes the Python package to TestPyPI. The workflow is triggered on push events. Additionally, the `python-publish.yml` workflow now depends on the `test` job. 

The version number of the package is now dynamically retrieved from `pyproject.toml` using Tomllib library in the `__init__.py` file.

Closes #123

* Update path to pyproject.toml file

The code changes update the path to the pyproject.toml file in the sqlalchemy_crud package. The new path is now located in the parent directory of the current file.

* Refactor workflows for publishing Python packages

- Removed unnecessary steps from the `poetry-pytest.yml` workflow.

* Refactor code to use `toml` library instead of `tomllib`

- Remove unnecessary import statement for `tomllib`
- Replace usage of `tomllib.load()` with `toml.load()`
- Update the import statement for the `toml` library

* Update GitHub workflow and project version

- Removed the "needs: [test]" line from the GitHub workflow file.
- Updated the project version in pyproject.toml from "v0.1.0-beta" to "v0.1.1-beta".

* Update python-publish.yml to comment out the pypi_token line.

- Commented out the pypi_token line in python-publish.yml.

* Update pypi_token in python-publish.yml

The commit updates the pypi_token in the python-publish.yml file to use the PYPI_TOKEN secret instead of GITHUB_TOKEN. This change ensures that the correct token is used for publishing the Python poetry package.

* Update Python publishing workflow

- Commented out the repository name and URL for testpypi