Update click package version to 8.1.5 and SQLAlchemy version to >=1.3.19, <2.0.0
This commit updates the click package version from 8.1.4 to 8.1.5 in the poetry.lock file, along with its associated hashes for the wheel and tar.gz files.

Additionally, it modifies the SQLAlchemy version constraint in the pyproject.toml file to be >=1.3.19 and <2.0.0.

These changes ensure that we are using the latest versions of these packages while maintaining compatibility with our project's requirements.