Bump version to 0.2.1 in pyproject.toml (#12)
* Bump version to 0.2.1 in pyproject.toml

This commit updates the version number in the pyproject.toml file from 0.2.0 to 0.2.1, reflecting a minor update to the project.

* Add test artifact upload step

This commit adds a new step to the workflow that uploads the test artifact "test-report.xml" using the EnricoMi/publish-unit-test-result-action@v2 action. This step is executed regardless of the outcome of previous steps.

* Update workflow to dynamically set Python versions and add steps to upload and publish test artifacts.

- Dynamically set Python versions in the workflow using `${{vars.python-versions}}`.
- Added step to upload test artifact `test-report.xml` using `actions/upload-artifact@v3`.
- Added new job "Publish" with steps to download test artifact and publish it using `EnricoMi/publish-unit-test-result-action@v2`.

* Update GitHub Actions workflow to publish test results

The commit updates the `.github/workflows/poetry-pytest.yml` file. It renames the `Publish` job to `Publish-Test`. This change ensures that the workflow publishes test results correctly.

* Update GitHub Actions workflow to include multiple Python versions for testing.

- Add an environment variable `python-versions` with a list of supported Python versions.
- Update the `matrix` strategy to use the `env.python-versions` variable.
- Modify the steps in both jobs to use the updated `python-version` matrix value.

* Update Python versions in GitHub Actions workflow

- Update the Python versions used in the GitHub Actions workflow from `env.python-versions` to `["3.8", "3.9", "3.10", "3.11"]`.
- This change ensures that the workflow runs on the specified Python versions for both jobs.

* Add check_name to Unit Tests for python ${{ matrix.python-version }}

* Update GitHub Actions workflow for code coverage and test artifacts

- Added step to upload coverage reports to Codecov
- Added step to upload coverage artifact
- Added step to upload test artifact
- Updated job dependencies and conditions

* Fix typo in pytest action file.

* feat: Update Python version in workflow

Updated the Python version in the workflow file to use the value specified in the matrix. This ensures that the correct Python version is set up for running tests.

* Update download artifact action to v3 in poetry-pytest.yml

The commit updates the download artifact action from v2 to v3 in the poetry-pytest.yml file. This change ensures compatibility with the latest version of the action and improves functionality.